requirements.txt
numpy == 1.18.5
matplotlib == 3.3.2
tensorflow == 2.3.1
Keras == 2.4.3
opencv-python
gTTS
pyspellchecker
